Create and Complete TAPP5
Will update later to add the capability to work with
pay that is hourly, only given as a range from indeed,
not given, or other cases.
I will also be updating to allow reading in a list of
job urls from a file.
Much later I may add the ability to take job postings
from other sites, like LinkedIn.